Fixed parents for blocks from Expr\ErrorSuppress
I encountered a bug where the Parser was not correctly resolving a
variable when I used an error suppress expression. I fixed this by
adding a parent to each of the two blocks introduced in
parseExpr_ErrorSuppress. The parent in each case is the block that just
added a Jump to the new block.

A parent for the block was eventually being added by the Simplifier
visitor, but this was not in time for variable resolution.